requirements.txt
tensorflow==2.12.0
pandas
gdown
dvc
mlflow==2.2.2
notebook==6.5.6
numpy
matplotlib
seaborn
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
scipy
Flask
Flask-Cors
pywinpty==2.0.13
-e .
